In progress: Making arg parsing more robust again
I forgot to make sure args were separated with commas. This could
be done with regex but then it would need to match \n, which we
want to leave for our own newline-eating code for the sake of
counting the types of newlines.

Now debugging named arg with unterminated quote getting picked up
by unnamed args regex